Feat: 텍스트 검색 기능  초안 구현 완료 (#97) - 일부 dto 필드타입 수정 - todo: 더미데이터로 쿼리 확인

+++ src/main/java/com/example/jejugudgo/domain/course/olle/dto/response/JejuOlleCourseResponseDetail.java
@@ -9,7 +9,7 @@ public record JejuOlleCourseResponseDetail(
         String endSpot,
         String totalTime,
         double starAvg,
-        Long reviewCount,
+        int reviewCount,
         // 태그 리스폰스 (tagResponse)
 
         // 스팟 리스폰스 (spotReseponse)
